Substitute variables in Makefiles
This change allows textual replacements of variables and
addsuffix/addprefix patterns (at least those present in v4.15, it
might not be generally applicable).

Furthermore, we default to x86 as the architecture now and
allow more than one parameter to 'include'.

This mainly fixes the problems we had with AMDs driver core with
the Makefiles beginning from drivers/gpu/amd/amdgpu.